Add ExampleStory
