+++ app/main.js
@@ -16,7 +16,7 @@
   \************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ AvO)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_physics__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_story__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/story */ \"./src/avo/story/index.js\");\n/* harmony import */ var _avo_interaction__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/interaction */ \"./src/avo/interaction/index.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\n\n\n\n\nvar searchParams = new URLSearchParams(window.location.search);\nvar DEBUG = searchParams.get('debug') || false;\nvar AvO = /*#__PURE__*/function () {\n  function AvO() {\n    var args = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n    _classCallCheck(this, AvO);\n    var _args$story = args.story,\n      story = _args$story === void 0 ? _avo_story__WEBPACK_IMPORTED_MODULE_2__[\"default\"] : _args$story,\n      _args$width = args.width,\n      width = _args$width === void 0 ? 24 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$width,\n      _args$height = args.height,\n      height = _args$height === void 0 ? 16 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$height;\n    this.html = {\n      main: document.getElementById('main'),\n      canvas: document.getElementById('canvas'),\n      homeMenu: document.getElementById('home-menu'),\n      interactionMenu: document.getElementById('interaction-menu'),\n      buttonHome: document.getElementById('button-home'),\n      buttonFullscreen: document.getElementById('button-fullscreen'),\n      buttonReload: document.getElementById('button-reload')\n    };\n    this.homeMenu = false;\n    this.setHomeMenu(false);\n    this.interactionMenu = false;\n    this.setInteractionMenu(false);\n    this.canvas2d = this.html.canvas.getContext('2d');\n    this.canvasWidth = width;\n    this.canvasHeight = height;\n    this._canvasHasCameraTransforms = false; // Safety check\n\n    this.camera = {\n      target: null,\n      // Target entity to follow. If null, camera is static.\n      x: 0,\n      y: 0,\n      zoom: 1\n    };\n    this.setupUI();\n    this.initialised = false;\n    this.hero = null;\n    this.entities = [];\n    this.rules = {};\n    this.story = new story(this);\n    this.assets = this.story.assets || {};\n    this.secretAssets = {};\n    this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n    this.playerInput = {\n      // Mouse/touchscreen input\n      pointerStart: undefined,\n      pointerCurrent: undefined,\n      pointerEnd: undefined,\n      // Keys that are currently being pressed.\n      // keysPressed = { key: { duration, acknowledged } }\n      keysPressed: {}\n    };\n    this.timeAccumulator = 0;\n    this.prevTime = null;\n    this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n  }\n  _createClass(AvO, [{\n    key: \"initialisationCheck\",\n    value: function initialisationCheck() {\n      var _this = this;\n      // Assets check\n      var allAssetsReady = true;\n      var numReadyAssets = 0;\n      var numTotalAssets = 0;\n      Object.keys(this.assets).forEach(function (id) {\n        var asset = _this.assets[id];\n        allAssetsReady = allAssetsReady && asset.ready;\n        if (asset.ready) numReadyAssets++;\n        numTotalAssets++;\n      });\n      Object.keys(this.secretAssets).forEach(function (id) {\n        var secretAsset = _this.secretAssets[id];\n        var secretAssetIsReady = secretAsset.ready || secretAsset.error;\n        allAssetsReady = allAssetsReady && secretAssetIsReady;\n        if (secretAssetIsReady) numReadyAssets++;\n        numTotalAssets++;\n      });\n\n      // Paint status\n      this.canvas2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      this.canvas2d.textAlign = 'start';\n      this.canvas2d.textBaseline = 'top';\n      this.canvas2d.fillStyle = '#ccc';\n      this.canvas2d.font = \"1em monospace\";\n      this.canvas2d.fillText(\"Loading \".concat(numReadyAssets, \" / \").concat(numTotalAssets, \" \"), _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n      if (allAssetsReady) {\n        // Clean up secret assets\n        Object.keys(this.secretAssets).forEach(function (id) {\n          if (_this.secretAssets[id].error) delete _this.secretAssets[id];\n        });\n\n        // Let's go!\n        this.initialised = true;\n        this.showUI();\n        this.story.start();\n      }\n    }\n\n    /*\r\n    Section: General Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    The main loop. Run a single frame of gameplay.\r\n    - time: the current/total time (milliseconds) since the game started.\r\n     */\n  }, {\n    key: \"main\",\n    value: function main(time) {\n      var timeStep = this.prevTime ? time - this.prevTime : time;\n      this.prevTime = time;\n      this.timeAccumulator += timeStep;\n      if (this.initialised) {\n        // Keep a consistent \"frame rate\" for logic processing\n        while (this.timeAccumulator >= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.EXPECTED_TIMESTEP) {\n          this.play(_avo_constants__WEBPACK_IMPORTED_MODULE_0__.EXPECTED_TIMESTEP);\n          this.timeAccumulator -= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.EXPECTED_TIMESTEP;\n        }\n        // Paint whenever possible\n        this.paint();\n      } else {\n        this.initialisationCheck();\n      }\n      this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n    }\n\n    /*\r\n    Run the gameplay/physics logic for a single frame.\r\n    - timeStep: the time (milliseconds) since the last frame.\r\n      We expect 60 frames per second.\r\n     */\n  }, {\n    key: \"play\",\n    value: function play(timeStep) {\n      var _this2 = this;\n      // If a menu is open, pause all action gameplay\n      if (this.homeMenu || this.interactionMenu) return;\n\n      // Run the action gameplay\n      // ----------------\n      for (var id in this.rules) {\n        this.rules[id].play(timeStep);\n      }\n      this.entities.forEach(function (entity) {\n        return entity.play(timeStep);\n      });\n      this.checkCollisions(timeStep);\n\n      // Cleanup\n      this.entities = this.entities.filter(function (entity) {\n        return !entity._expired;\n      });\n      for (var _id in this.rules) {\n        if (this.rules[_id]._expired) delete this.rules[_id];\n      }\n\n      // Sort Entities along the y-axis, for paint()/rendering purposes.\n      // WARNING: inefficient\n      this.entities.sort(function (a, b) {\n        return a.y - b.y;\n      });\n      // ----------------\n\n      // Increment the duration of each currently pressed key\n      Object.keys(this.playerInput.keysPressed).forEach(function (key) {\n        if (_this2.playerInput.keysPressed[key]) _this2.playerInput.keysPressed[key].duration += timeStep;\n      });\n    }\n\n    /*\r\n    Paint/draw the game visuals onto the canvas.\r\n     */\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var _this3 = this;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n\n      // Camera Controls: focus the camera on the target entity, if any.\n      // ----------------\n      if (camera.target) {\n        camera.x = this.canvasWidth / 2 - camera.target.x * camera.zoom;\n        camera.y = this.canvasHeight / 2 - camera.target.y * camera.zoom;\n      }\n      c2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      c2d.resetTransform();\n      c2d.strokeStyle = 'rgba(128, 128, 128, 0.05)';\n      c2d.lineWidth = 2;\n      // ----------------\n\n      // Draw grid\n      // ----------------\n      var GRID_SIZE = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * camera.zoom;\n      var offsetX = this.camera.x % GRID_SIZE - GRID_SIZE;\n      var offsetY = this.camera.y % GRID_SIZE - GRID_SIZE;\n      for (var y = offsetY; y < this.canvasHeight; y += GRID_SIZE) {\n        for (var x = offsetX; x < this.canvasWidth; x += GRID_SIZE) {\n          c2d.beginPath();\n          c2d.rect(x, y, GRID_SIZE, GRID_SIZE);\n          c2d.stroke();\n\n          // Debug Grid\n          if (DEBUG) {\n            c2d.fillStyle = '#ccc';\n            c2d.font = \"\".concat(camera.zoom * 0.5, \"em Source Code Pro\");\n            c2d.textAlign = 'center';\n            c2d.textBaseline = 'middle';\n            var col = Math.floor((x - this.camera.x) / GRID_SIZE);\n            var row = Math.floor((y - this.camera.y) / GRID_SIZE);\n            c2d.fillText(col + ',' + row, x + GRID_SIZE / 2, y + GRID_SIZE / 2); // using template strings here messes up colours in Brackets.\n          }\n        }\n      }\n      // ----------------\n\n      // Draw entities and other elements\n      // ----------------\n      var _loop = function _loop(layer) {\n        _this3.entities.forEach(function (entity) {\n          return entity.paint(layer);\n        });\n        for (var id in _this3.rules) {\n          _this3.rules[id].paint(layer);\n        }\n      };\n      for (var layer = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MIN_LAYER; layer <= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MAX_LAYER; layer++) {\n        _loop(layer);\n      }\n      // ----------------\n\n      // Draw player input\n      // ----------------\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN && this.hero && this.playerInput.pointerCurrent) {\n        var inputCoords = this.playerInput.pointerCurrent;\n        c2d.strokeStyle = '#888';\n        c2d.lineWidth = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 8;\n        c2d.beginPath();\n        c2d.arc(inputCoords.x, inputCoords.y, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, 0, 2 * Math.PI);\n        c2d.stroke();\n      }\n      // ----------------\n    }\n\n    /*\r\n    Section: UI and Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"setupUI\",\n    value: function setupUI() {\n      this.html.canvas.width = this.canvasWidth;\n      this.html.canvas.height = this.canvasHeight;\n      this.canvas2d.imageSmoothingEnabled = false; /* Pixel art: Maintains sprites' pixel sharpness when scaled up via drawImage() */\n\n      if (window.PointerEvent) {\n        this.html.canvas.addEventListener('pointerdown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('pointermove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('pointerup', this.onPointerUp.bind(this));\n        this.html.canvas.addEventListener('pointercancel', this.onPointerUp.bind(this));\n      } else {\n        this.html.canvas.addEventListener('mousedown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('mousemove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('mouseup', this.onPointerUp.bind(this));\n      }\n\n      // Prevent \"touch and hold to open context menu\" menu on touchscreens.\n      this.html.canvas.addEventListener('touchstart', stopEvent);\n      this.html.canvas.addEventListener('touchmove', stopEvent);\n      this.html.canvas.addEventListener('touchend', stopEvent);\n      this.html.canvas.addEventListener('touchcancel', stopEvent);\n      this.html.buttonHome.addEventListener('click', this.buttonHome_onClick.bind(this));\n      this.html.buttonFullscreen.addEventListener('click', this.buttonFullscreen_onClick.bind(this));\n      this.html.buttonReload.addEventListener('click', this.buttonReload_onClick.bind(this));\n      this.html.main.addEventListener('keydown', this.onKeyDown.bind(this));\n      this.html.main.addEventListener('keyup', this.onKeyUp.bind(this));\n      window.addEventListener('resize', this.updateUI.bind(this));\n      this.updateUI();\n      this.hideUI(); // Hide until all assets are ready\n\n      this.html.main.focus();\n    }\n  }, {\n    key: \"hideUI\",\n    value: function hideUI() {\n      this.html.buttonHome.style.visibility = 'hidden';\n      this.html.buttonReload.style.visibility = 'hidden';\n    }\n  }, {\n    key: \"showUI\",\n    value: function showUI() {\n      this.html.buttonHome.style.visibility = 'visible';\n      this.html.buttonReload.style.visibility = 'visible';\n    }\n  }, {\n    key: \"updateUI\",\n    value: function updateUI() {\n      // Fit the interaction layers (menus, etc) to the canvas\n      var mainDivBounds = this.html.main.getBoundingClientRect();\n      var canvasBounds = this.html.canvas.getBoundingClientRect();\n      this.html.homeMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.homeMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.homeMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.homeMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n      this.html.interactionMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.interactionMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.interactionMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.interactionMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n    }\n  }, {\n    key: \"setHomeMenu\",\n    value: function setHomeMenu(homeMenu) {\n      this.homeMenu = homeMenu;\n      if (homeMenu) {\n        this.html.homeMenu.style.visibility = 'visible';\n        this.html.buttonReload.style.visibility = 'hidden';\n      } else {\n        this.html.homeMenu.style.visibility = 'hidden';\n        this.html.buttonReload.style.visibility = 'visible';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"setInteractionMenu\",\n    value: function setInteractionMenu(interactionMenu) {\n      var div = this.html.interactionMenu;\n      this.interactionMenu && this.interactionMenu.unload(); // Unload the old menu, if any\n      this.interactionMenu = interactionMenu; // Set the new menu\n\n      if (interactionMenu) {\n        while (div.firstChild) {\n          div.removeChild(div.firstChild);\n        } // Clear div\n        interactionMenu.load(div); // load the new menu\n        div.style.visibility = 'visible';\n      } else {\n        div.style.visibility = 'hidden';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"onPointerDown\",\n    value: function onPointerDown(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN;\n      this.playerInput.pointerStart = coords;\n      this.playerInput.pointerCurrent = coords;\n      this.playerInput.pointerEnd = undefined;\n      this.html.main.focus();\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerMove\",\n    value: function onPointerMove(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerInput.pointerCurrent = coords;\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerUp\",\n    value: function onPointerUp(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN) {\n        this.playerInput.pointerEnd = coords;\n        this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      }\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onKeyDown\",\n    value: function onKeyDown(e) {\n      var _this$hero, _this$hero2;\n      // Special cases\n      switch (e.key) {\n        // Open home menu\n        case 'Escape':\n          this.setHomeMenu(!this.homeMenu);\n          break;\n\n        // DEBUG\n        case 'z':\n          if (!this.interactionMenu) {\n            this.setInteractionMenu(new _avo_interaction__WEBPACK_IMPORTED_MODULE_3__[\"default\"](this));\n          }\n          break;\n\n        // DEBUG\n        case 'c':\n          if (((_this$hero = this.hero) === null || _this$hero === void 0 ? void 0 : _this$hero.spriteStyle) === 'toon') {\n            this.hero.spriteStyle = 'zelda';\n          } else if (((_this$hero2 = this.hero) === null || _this$hero2 === void 0 ? void 0 : _this$hero2.spriteStyle) === 'zelda') {\n            this.hero.spriteStyle = 'toon';\n          }\n          break;\n        case '-':\n        case '_':\n          this.camera.zoom = Math.max(0.5, this.camera.zoom - 0.5);\n          break;\n        case '+':\n        case '=':\n          this.camera.zoom = Math.min(4, this.camera.zoom + 0.5);\n          break;\n      }\n\n      // General input\n      if (!this.playerInput.keysPressed[e.key]) {\n        this.playerInput.keysPressed[e.key] = {\n          duration: 0,\n          acknowledged: false\n        };\n      }\n    }\n  }, {\n    key: \"onKeyUp\",\n    value: function onKeyUp(e) {\n      this.playerInput.keysPressed[e.key] = undefined;\n    }\n  }, {\n    key: \"buttonHome_onClick\",\n    value: function buttonHome_onClick() {\n      this.setHomeMenu(!this.homeMenu);\n    }\n  }, {\n    key: \"buttonFullscreen_onClick\",\n    value: function buttonFullscreen_onClick() {\n      var _this4 = this;\n      var isFullscreen = document.fullscreenElement;\n      if (!isFullscreen) {\n        if (this.html.main.requestFullscreen) {\n          this.html.main.requestFullscreen().then(function () {\n            _this4.html.main.className = 'fullscreen';\n            _this4.updateUI();\n          })[\"catch\"](function (err) {\n            console.error('requestFullscreen() error: ', err);\n          });\n        }\n      } else {\n        var _document$exitFullscr, _document;\n        (_document$exitFullscr = (_document = document).exitFullscreen) === null || _document$exitFullscr === void 0 || _document$exitFullscr.call(_document).then(function () {\n          _this4.html.main.className = '';\n          _this4.updateUI();\n        })[\"catch\"](function (err) {\n          console.error('exitFullscreen() error: ', err);\n        });\n      }\n    }\n  }, {\n    key: \"buttonReload_onClick\",\n    value: function buttonReload_onClick() {\n      this.story.reload();\n    }\n\n    /*\r\n    Section: Gameplay\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"addEntity\",\n    value: function addEntity(entity) {\n      if (!entity) return null;\n      if (!this.entities.includes(entity)) this.entities.push(entity);\n      return entity;\n    }\n  }, {\n    key: \"removeEntity\",\n    value: function removeEntity(entityOrMatchingFn) {\n      if (!entityOrMatchingFn) return;\n      if (typeof entityOrMatchingFn === 'function') {\n        this.entities.filter(entityOrMatchingFn).forEach(function (entity) {\n          entity._expired = true;\n        });\n      } else if (this.entities.includes(entityOrMatchingFn)) {\n        entityOrMatchingFn._expired = true;\n      }\n    }\n  }, {\n    key: \"addRule\",\n    value: function addRule(rule) {\n      if (!rule) return;\n      var id = rule._type;\n      this.rules[id] = rule;\n    }\n  }, {\n    key: \"clearRules\",\n    value: function clearRules() {\n      for (var id in this.rules) {\n        delete this.rules[id];\n      }\n    }\n\n    /*\r\n    Section: Painting\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Applies camera transforms to the canvas.\r\n    Should be run right before drawing an Entity (or etc) so the object is drawn\r\n    relative to the camera's view.\r\n     */\n  }, {\n    key: \"applyCameraTransforms\",\n    value: function applyCameraTransforms() {\n      if (this._canvasHasCameraTransforms) throw new Error('Canvas already has camera transforms.');\n      this._canvasHasCameraTransforms = true;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n      c2d.save();\n      c2d.translate(camera.x, camera.y);\n      c2d.scale(camera.zoom, camera.zoom);\n    }\n\n    /*\r\n    Removes camera transforms from the canvas.\r\n     */\n  }, {\n    key: \"undoCameraTransforms\",\n    value: function undoCameraTransforms() {\n      if (!this._canvasHasCameraTransforms) throw new Error('Canvas doesn\\'t have camera transforms.');\n      this._canvasHasCameraTransforms = false;\n      this.canvas2d.restore();\n    }\n\n    /*\r\n    Section: Misc\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"checkCollisions\",\n    value: function checkCollisions(timeStep) {\n      for (var a = 0; a < this.entities.length; a++) {\n        var entityA = this.entities[a];\n        for (var b = a + 1; b < this.entities.length; b++) {\n          var entityB = this.entities[b];\n          var collisionCorrection = _avo_physics__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(entityA, entityB);\n          if (collisionCorrection) {\n            entityA.onCollision(entityB, collisionCorrection.a);\n            entityB.onCollision(entityA, collisionCorrection.b);\n          }\n        }\n      }\n    }\n  }]);\n  return AvO;\n}();\n\nfunction getEventCoords(event, element) {\n  var xRatio = element.width && element.offsetWidth ? element.width / element.offsetWidth : 1;\n  var yRatio = element.height && element.offsetHeight ? element.height / element.offsetHeight : 1;\n  var x = event.offsetX * xRatio;\n  var y = event.offsetY * yRatio;\n  return {\n    x: x,\n    y: y\n  };\n}\nfunction stopEvent(e) {\n  if (!e) return false;\n  e.preventDefault && e.preventDefault();\n  e.stopPropagation && e.stopPropagation();\n  e.returnValue = false;\n  e.cancelBubble = true;\n  return false;\n}\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/avo.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ AvO)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_physics__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_story_types_example_story__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/story/types/example-story */ \"./src/avo/story/types/example-story.js\");\n/* harmony import */ var _avo_interaction__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/interaction */ \"./src/avo/interaction/index.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\n\n\n\n\nvar searchParams = new URLSearchParams(window.location.search);\nvar DEBUG = searchParams.get('debug') || false;\nvar AvO = /*#__PURE__*/function () {\n  function AvO() {\n    var args = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n    _classCallCheck(this, AvO);\n    var _args$story = args.story,\n      story = _args$story === void 0 ? _avo_story_types_example_story__WEBPACK_IMPORTED_MODULE_2__[\"default\"] : _args$story,\n      _args$width = args.width,\n      width = _args$width === void 0 ? 24 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$width,\n      _args$height = args.height,\n      height = _args$height === void 0 ? 16 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$height;\n    this.html = {\n      main: document.getElementById('main'),\n      canvas: document.getElementById('canvas'),\n      homeMenu: document.getElementById('home-menu'),\n      interactionMenu: document.getElementById('interaction-menu'),\n      buttonHome: document.getElementById('button-home'),\n      buttonFullscreen: document.getElementById('button-fullscreen'),\n      buttonReload: document.getElementById('button-reload')\n    };\n    this.homeMenu = false;\n    this.setHomeMenu(false);\n    this.interactionMenu = false;\n    this.setInteractionMenu(false);\n    this.canvas2d = this.html.canvas.getContext('2d');\n    this.canvasWidth = width;\n    this.canvasHeight = height;\n    this._canvasHasCameraTransforms = false; // Safety check\n\n    this.camera = {\n      target: null,\n      // Target entity to follow. If null, camera is static.\n      x: 0,\n      y: 0,\n      zoom: 1\n    };\n    this.setupUI();\n    this.initialised = false;\n    this.hero = null;\n    this.entities = [];\n    this.rules = {};\n    this.story = new story(this);\n    this.assets = this.story.assets || {};\n    this.secretAssets = {};\n    this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n    this.playerInput = {\n      // Mouse/touchscreen input\n      pointerStart: undefined,\n      pointerCurrent: undefined,\n      pointerEnd: undefined,\n      // Keys that are currently being pressed.\n      // keysPressed = { key: { duration, acknowledged } }\n      keysPressed: {}\n    };\n    this.timeAccumulator = 0;\n    this.prevTime = null;\n    this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n  }\n  _createClass(AvO, [{\n    key: \"initialisationCheck\",\n    value: function initialisationCheck() {\n      var _this = this;\n      // Assets check\n      var allAssetsReady = true;\n      var numReadyAssets = 0;\n      var numTotalAssets = 0;\n      Object.keys(this.assets).forEach(function (id) {\n        var asset = _this.assets[id];\n        allAssetsReady = allAssetsReady && asset.ready;\n        if (asset.ready) numReadyAssets++;\n        numTotalAssets++;\n      });\n      Object.keys(this.secretAssets).forEach(function (id) {\n        var secretAsset = _this.secretAssets[id];\n        var secretAssetIsReady = secretAsset.ready || secretAsset.error;\n        allAssetsReady = allAssetsReady && secretAssetIsReady;\n        if (secretAssetIsReady) numReadyAssets++;\n        numTotalAssets++;\n      });\n\n      // Paint status\n      this.canvas2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      this.canvas2d.textAlign = 'start';\n      this.canvas2d.textBaseline = 'top';\n      this.canvas2d.fillStyle = '#ccc';\n      this.canvas2d.font = \"1em monospace\";\n      this.canvas2d.fillText(\"Loading \".concat(numReadyAssets, \" / \").concat(numTotalAssets, \" \"), _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n      if (allAssetsReady) {\n        // Clean up secret assets\n        Object.keys(this.secretAssets).forEach(function (id) {\n          if (_this.secretAssets[id].error) delete _this.secretAssets[id];\n        });\n\n        // Let's go!\n        this.initialised = true;\n        this.showUI();\n        this.story.start();\n      }\n    }\n\n    /*\r\n    Section: General Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    The main loop. Run a single frame of gameplay.\r\n    - time: the current/total time (milliseconds) since the game started.\r\n     */\n  }, {\n    key: \"main\",\n    value: function main(time) {\n      var timeStep = this.prevTime ? time - this.prevTime : time;\n      this.prevTime = time;\n      this.timeAccumulator += timeStep;\n      if (this.initialised) {\n        // Keep a consistent \"frame rate\" for logic processing\n        while (this.timeAccumulator >= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.EXPECTED_TIMESTEP) {\n          this.play(_avo_constants__WEBPACK_IMPORTED_MODULE_0__.EXPECTED_TIMESTEP);\n          this.timeAccumulator -= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.EXPECTED_TIMESTEP;\n        }\n        // Paint whenever possible\n        this.paint();\n      } else {\n        this.initialisationCheck();\n      }\n      this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n    }\n\n    /*\r\n    Run the gameplay/physics logic for a single frame.\r\n    - timeStep: the time (milliseconds) since the last frame.\r\n      We expect 60 frames per second.\r\n     */\n  }, {\n    key: \"play\",\n    value: function play(timeStep) {\n      var _this2 = this;\n      // If a menu is open, pause all action gameplay\n      if (this.homeMenu || this.interactionMenu) return;\n\n      // Run the action gameplay\n      // ----------------\n      for (var id in this.rules) {\n        this.rules[id].play(timeStep);\n      }\n      this.entities.forEach(function (entity) {\n        return entity.play(timeStep);\n      });\n      this.checkCollisions(timeStep);\n\n      // Cleanup\n      this.entities = this.entities.filter(function (entity) {\n        return !entity._expired;\n      });\n      for (var _id in this.rules) {\n        if (this.rules[_id]._expired) delete this.rules[_id];\n      }\n\n      // Sort Entities along the y-axis, for paint()/rendering purposes.\n      // WARNING: inefficient\n      this.entities.sort(function (a, b) {\n        return a.y - b.y;\n      });\n      // ----------------\n\n      // Increment the duration of each currently pressed key\n      Object.keys(this.playerInput.keysPressed).forEach(function (key) {\n        if (_this2.playerInput.keysPressed[key]) _this2.playerInput.keysPressed[key].duration += timeStep;\n      });\n    }\n\n    /*\r\n    Paint/draw the game visuals onto the canvas.\r\n     */\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var _this3 = this;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n\n      // Camera Controls: focus the camera on the target entity, if any.\n      // ----------------\n      if (camera.target) {\n        camera.x = this.canvasWidth / 2 - camera.target.x * camera.zoom;\n        camera.y = this.canvasHeight / 2 - camera.target.y * camera.zoom;\n      }\n      c2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      c2d.resetTransform();\n      c2d.strokeStyle = 'rgba(128, 128, 128, 0.05)';\n      c2d.lineWidth = 2;\n      // ----------------\n\n      // Draw grid\n      // ----------------\n      var GRID_SIZE = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * camera.zoom;\n      var offsetX = this.camera.x % GRID_SIZE - GRID_SIZE;\n      var offsetY = this.camera.y % GRID_SIZE - GRID_SIZE;\n      for (var y = offsetY; y < this.canvasHeight; y += GRID_SIZE) {\n        for (var x = offsetX; x < this.canvasWidth; x += GRID_SIZE) {\n          c2d.beginPath();\n          c2d.rect(x, y, GRID_SIZE, GRID_SIZE);\n          c2d.stroke();\n\n          // Debug Grid\n          if (DEBUG) {\n            c2d.fillStyle = '#ccc';\n            c2d.font = \"\".concat(camera.zoom * 0.5, \"em Source Code Pro\");\n            c2d.textAlign = 'center';\n            c2d.textBaseline = 'middle';\n            var col = Math.floor((x - this.camera.x) / GRID_SIZE);\n            var row = Math.floor((y - this.camera.y) / GRID_SIZE);\n            c2d.fillText(col + ',' + row, x + GRID_SIZE / 2, y + GRID_SIZE / 2); // using template strings here messes up colours in Brackets.\n          }\n        }\n      }\n      // ----------------\n\n      // Draw entities and other elements\n      // ----------------\n      var _loop = function _loop(layer) {\n        _this3.entities.forEach(function (entity) {\n          return entity.paint(layer);\n        });\n        for (var id in _this3.rules) {\n          _this3.rules[id].paint(layer);\n        }\n      };\n      for (var layer = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MIN_LAYER; layer <= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MAX_LAYER; layer++) {\n        _loop(layer);\n      }\n      // ----------------\n\n      // Draw player input\n      // ----------------\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN && this.hero && this.playerInput.pointerCurrent) {\n        var inputCoords = this.playerInput.pointerCurrent;\n        c2d.strokeStyle = '#888';\n        c2d.lineWidth = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 8;\n        c2d.beginPath();\n        c2d.arc(inputCoords.x, inputCoords.y, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, 0, 2 * Math.PI);\n        c2d.stroke();\n      }\n      // ----------------\n    }\n\n    /*\r\n    Section: UI and Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"setupUI\",\n    value: function setupUI() {\n      this.html.canvas.width = this.canvasWidth;\n      this.html.canvas.height = this.canvasHeight;\n      this.canvas2d.imageSmoothingEnabled = false; /* Pixel art: Maintains sprites' pixel sharpness when scaled up via drawImage() */\n\n      if (window.PointerEvent) {\n        this.html.canvas.addEventListener('pointerdown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('pointermove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('pointerup', this.onPointerUp.bind(this));\n        this.html.canvas.addEventListener('pointercancel', this.onPointerUp.bind(this));\n      } else {\n        this.html.canvas.addEventListener('mousedown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('mousemove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('mouseup', this.onPointerUp.bind(this));\n      }\n\n      // Prevent \"touch and hold to open context menu\" menu on touchscreens.\n      this.html.canvas.addEventListener('touchstart', stopEvent);\n      this.html.canvas.addEventListener('touchmove', stopEvent);\n      this.html.canvas.addEventListener('touchend', stopEvent);\n      this.html.canvas.addEventListener('touchcancel', stopEvent);\n      this.html.buttonHome.addEventListener('click', this.buttonHome_onClick.bind(this));\n      this.html.buttonFullscreen.addEventListener('click', this.buttonFullscreen_onClick.bind(this));\n      this.html.buttonReload.addEventListener('click', this.buttonReload_onClick.bind(this));\n      this.html.main.addEventListener('keydown', this.onKeyDown.bind(this));\n      this.html.main.addEventListener('keyup', this.onKeyUp.bind(this));\n      window.addEventListener('resize', this.updateUI.bind(this));\n      this.updateUI();\n      this.hideUI(); // Hide until all assets are ready\n\n      this.html.main.focus();\n    }\n  }, {\n    key: \"hideUI\",\n    value: function hideUI() {\n      this.html.buttonHome.style.visibility = 'hidden';\n      this.html.buttonReload.style.visibility = 'hidden';\n    }\n  }, {\n    key: \"showUI\",\n    value: function showUI() {\n      this.html.buttonHome.style.visibility = 'visible';\n      this.html.buttonReload.style.visibility = 'visible';\n    }\n  }, {\n    key: \"updateUI\",\n    value: function updateUI() {\n      // Fit the interaction layers (menus, etc) to the canvas\n      var mainDivBounds = this.html.main.getBoundingClientRect();\n      var canvasBounds = this.html.canvas.getBoundingClientRect();\n      this.html.homeMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.homeMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.homeMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.homeMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n      this.html.interactionMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.interactionMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.interactionMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.interactionMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n    }\n  }, {\n    key: \"setHomeMenu\",\n    value: function setHomeMenu(homeMenu) {\n      this.homeMenu = homeMenu;\n      if (homeMenu) {\n        this.html.homeMenu.style.visibility = 'visible';\n        this.html.buttonReload.style.visibility = 'hidden';\n      } else {\n        this.html.homeMenu.style.visibility = 'hidden';\n        this.html.buttonReload.style.visibility = 'visible';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"setInteractionMenu\",\n    value: function setInteractionMenu(interactionMenu) {\n      var div = this.html.interactionMenu;\n      this.interactionMenu && this.interactionMenu.unload(); // Unload the old menu, if any\n      this.interactionMenu = interactionMenu; // Set the new menu\n\n      if (interactionMenu) {\n        while (div.firstChild) {\n          div.removeChild(div.firstChild);\n        } // Clear div\n        interactionMenu.load(div); // load the new menu\n        div.style.visibility = 'visible';\n      } else {\n        div.style.visibility = 'hidden';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"onPointerDown\",\n    value: function onPointerDown(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN;\n      this.playerInput.pointerStart = coords;\n      this.playerInput.pointerCurrent = coords;\n      this.playerInput.pointerEnd = undefined;\n      this.html.main.focus();\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerMove\",\n    value: function onPointerMove(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerInput.pointerCurrent = coords;\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerUp\",\n    value: function onPointerUp(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN) {\n        this.playerInput.pointerEnd = coords;\n        this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      }\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onKeyDown\",\n    value: function onKeyDown(e) {\n      var _this$hero, _this$hero2;\n      // Special cases\n      switch (e.key) {\n        // Open home menu\n        case 'Escape':\n          this.setHomeMenu(!this.homeMenu);\n          break;\n\n        // DEBUG\n        case 'z':\n          if (!this.interactionMenu) {\n            this.setInteractionMenu(new _avo_interaction__WEBPACK_IMPORTED_MODULE_3__[\"default\"](this));\n          }\n          break;\n\n        // DEBUG\n        case 'c':\n          if (((_this$hero = this.hero) === null || _this$hero === void 0 ? void 0 : _this$hero.spriteStyle) === 'toon') {\n            this.hero.spriteStyle = 'zelda';\n          } else if (((_this$hero2 = this.hero) === null || _this$hero2 === void 0 ? void 0 : _this$hero2.spriteStyle) === 'zelda') {\n            this.hero.spriteStyle = 'toon';\n          }\n          break;\n        case '-':\n        case '_':\n          this.camera.zoom = Math.max(0.5, this.camera.zoom - 0.5);\n          break;\n        case '+':\n        case '=':\n          this.camera.zoom = Math.min(4, this.camera.zoom + 0.5);\n          break;\n      }\n\n      // General input\n      if (!this.playerInput.keysPressed[e.key]) {\n        this.playerInput.keysPressed[e.key] = {\n          duration: 0,\n          acknowledged: false\n        };\n      }\n    }\n  }, {\n    key: \"onKeyUp\",\n    value: function onKeyUp(e) {\n      this.playerInput.keysPressed[e.key] = undefined;\n    }\n  }, {\n    key: \"buttonHome_onClick\",\n    value: function buttonHome_onClick() {\n      this.setHomeMenu(!this.homeMenu);\n    }\n  }, {\n    key: \"buttonFullscreen_onClick\",\n    value: function buttonFullscreen_onClick() {\n      var _this4 = this;\n      var isFullscreen = document.fullscreenElement;\n      if (!isFullscreen) {\n        if (this.html.main.requestFullscreen) {\n          this.html.main.requestFullscreen().then(function () {\n            _this4.html.main.className = 'fullscreen';\n            _this4.updateUI();\n          })[\"catch\"](function (err) {\n            console.error('requestFullscreen() error: ', err);\n          });\n        }\n      } else {\n        var _document$exitFullscr, _document;\n        (_document$exitFullscr = (_document = document).exitFullscreen) === null || _document$exitFullscr === void 0 || _document$exitFullscr.call(_document).then(function () {\n          _this4.html.main.className = '';\n          _this4.updateUI();\n        })[\"catch\"](function (err) {\n          console.error('exitFullscreen() error: ', err);\n        });\n      }\n    }\n  }, {\n    key: \"buttonReload_onClick\",\n    value: function buttonReload_onClick() {\n      this.story.reload();\n    }\n\n    /*\r\n    Section: Gameplay\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"addEntity\",\n    value: function addEntity(entity) {\n      if (!entity) return null;\n      if (!this.entities.includes(entity)) this.entities.push(entity);\n      return entity;\n    }\n  }, {\n    key: \"removeEntity\",\n    value: function removeEntity(entityOrMatchingFn) {\n      if (!entityOrMatchingFn) return;\n      if (typeof entityOrMatchingFn === 'function') {\n        this.entities.filter(entityOrMatchingFn).forEach(function (entity) {\n          entity._expired = true;\n        });\n      } else if (this.entities.includes(entityOrMatchingFn)) {\n        entityOrMatchingFn._expired = true;\n      }\n    }\n  }, {\n    key: \"addRule\",\n    value: function addRule(rule) {\n      if (!rule) return;\n      var id = rule._type;\n      this.rules[id] = rule;\n    }\n  }, {\n    key: \"clearRules\",\n    value: function clearRules() {\n      for (var id in this.rules) {\n        delete this.rules[id];\n      }\n    }\n\n    /*\r\n    Section: Painting\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Applies camera transforms to the canvas.\r\n    Should be run right before drawing an Entity (or etc) so the object is drawn\r\n    relative to the camera's view.\r\n     */\n  }, {\n    key: \"applyCameraTransforms\",\n    value: function applyCameraTransforms() {\n      if (this._canvasHasCameraTransforms) throw new Error('Canvas already has camera transforms.');\n      this._canvasHasCameraTransforms = true;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n      c2d.save();\n      c2d.translate(camera.x, camera.y);\n      c2d.scale(camera.zoom, camera.zoom);\n    }\n\n    /*\r\n    Removes camera transforms from the canvas.\r\n     */\n  }, {\n    key: \"undoCameraTransforms\",\n    value: function undoCameraTransforms() {\n      if (!this._canvasHasCameraTransforms) throw new Error('Canvas doesn\\'t have camera transforms.');\n      this._canvasHasCameraTransforms = false;\n      this.canvas2d.restore();\n    }\n\n    /*\r\n    Section: Misc\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"checkCollisions\",\n    value: function checkCollisions(timeStep) {\n      for (var a = 0; a < this.entities.length; a++) {\n        var entityA = this.entities[a];\n        for (var b = a + 1; b < this.entities.length; b++) {\n          var entityB = this.entities[b];\n          var collisionCorrection = _avo_physics__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(entityA, entityB);\n          if (collisionCorrection) {\n            entityA.onCollision(entityB, collisionCorrection.a);\n            entityB.onCollision(entityA, collisionCorrection.b);\n          }\n        }\n      }\n    }\n  }]);\n  return AvO;\n}();\n\nfunction getEventCoords(event, element) {\n  var xRatio = element.width && element.offsetWidth ? element.width / element.offsetWidth : 1;\n  var yRatio = element.height && element.offsetHeight ? element.height / element.offsetHeight : 1;\n  var x = event.offsetX * xRatio;\n  var y = event.offsetY * yRatio;\n  return {\n    x: x,\n    y: y\n  };\n}\nfunction stopEvent(e) {\n  if (!e) return false;\n  e.preventDefault && e.preventDefault();\n  e.stopPropagation && e.stopPropagation();\n  e.returnValue = false;\n  e.cancelBubble = true;\n  return false;\n}\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/avo.js?");
 
 /***/ }),
 
@@ -190,23 +190,23 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/avo/story/index.js":
-/*!********************************!*\
-  !*** ./src/avo/story/index.js ***!
-  \********************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _story__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./story */ \"./src/avo/story/story.js\");\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (_story__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/index.js?");
-
-/***/ }),
-
 /***/ "./src/avo/story/story.js":
 /*!********************************!*\
   !*** ./src/avo/story/story.js ***!
   \********************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Story)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_image_asset__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/image-asset */ \"./src/avo/image-asset.js\");\n/* harmony import */ var _avo_entity_types_hero__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/entity/types/hero */ \"./src/avo/entity/types/hero.js\");\n/* harmony import */ var _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/entity/types/wall */ \"./src/avo/entity/types/wall.js\");\n/* harmony import */ var _avo_entity_types_ball__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/entity/types/ball */ \"./src/avo/entity/types/ball.js\");\n/* harmony import */ var _avo_entity_types_enemy__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/entity/types/enemy */ \"./src/avo/entity/types/enemy.js\");\n/* harmony import */ var _avo_rule_types_zelda_controls__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! @avo/rule/types/zelda-controls */ \"./src/avo/rule/types/zelda-controls.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\n\n\n\n\n\n\n\nvar Story = /*#__PURE__*/function () {\n  function Story(app) {\n    _classCallCheck(this, Story);\n    this._app = app;\n  }\n  _createClass(Story, [{\n    key: \"assets\",\n    get: function get() {\n      return {\n        \"hero-4dir\": new _avo_image_asset__WEBPACK_IMPORTED_MODULE_1__[\"default\"]('assets/avo-sprites-2022-05-samiel.png'),\n        \"hero-2dir\": new _avo_image_asset__WEBPACK_IMPORTED_MODULE_1__[\"default\"]('assets/avo-sprites-2022-10-samiel-2dir.png')\n      };\n    }\n  }, {\n    key: \"start\",\n    value: function start() {\n      this.reset();\n      this.generate_zelda_default();\n    }\n  }, {\n    key: \"reset\",\n    value: function reset() {\n      var app = this._app;\n      app.hero = undefined;\n      app.entities = [];\n      app.clearRules();\n      app.camera.target = null;\n      app.camera.x = 0;\n      app.camera.y = 0;\n      app.camera.zoom = 1;\n      app.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      app.setInteractionMenu(false);\n    }\n  }, {\n    key: \"reload\",\n    value: function reload() {\n      this.start();\n    }\n\n    /*\r\n    Default top-down adventure level.\r\n     */\n  }, {\n    key: \"generate_zelda_default\",\n    value: function generate_zelda_default() {\n      var app = this._app;\n      app.hero = app.addEntity(new _avo_entity_types_hero__WEBPACK_IMPORTED_MODULE_2__[\"default\"](app, 11, 1));\n      app.camera.target = app.hero;\n      app.addRule(new _avo_rule_types_zelda_controls__WEBPACK_IMPORTED_MODULE_6__[\"default\"](app));\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 0, 0, 1, 23)); // West Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 22, 0, 1, 23)); // East Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 1, 0, 21, 1)); // North Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 1, 22, 21, 1)); // South Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 3, 2, 3, 1));\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 3, 4, 3, 1));\n      app.addEntity(new _avo_entity_types_ball__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 8, 6));\n      var enemy = app.addEntity(new _avo_entity_types_enemy__WEBPACK_IMPORTED_MODULE_5__[\"default\"](app, 4, 8));\n      enemy.rotation = -45 / 180 * Math.PI;\n    }\n  }]);\n  return Story;\n}();\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/story.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Story)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\n\nvar Story = /*#__PURE__*/function () {\n  function Story(app) {\n    _classCallCheck(this, Story);\n    this._app = app;\n  }\n  _createClass(Story, [{\n    key: \"assets\",\n    get: function get() {\n      return {};\n    }\n  }, {\n    key: \"start\",\n    value: function start() {\n      this.reset();\n    }\n  }, {\n    key: \"reset\",\n    value: function reset() {\n      var app = this._app;\n      app.hero = undefined;\n      app.entities = [];\n      app.clearRules();\n      app.camera.target = null;\n      app.camera.x = 0;\n      app.camera.y = 0;\n      app.camera.zoom = 1;\n      app.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      app.setInteractionMenu(false);\n    }\n  }, {\n    key: \"reload\",\n    value: function reload() {\n      this.start();\n    }\n  }]);\n  return Story;\n}();\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/story.js?");
+
+/***/ }),
+
+/***/ "./src/avo/story/types/example-story.js":
+/*!**********************************************!*\
+  !*** ./src/avo/story/types/example-story.js ***!
+  \**********************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ExampleStory)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _story__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../story */ \"./src/avo/story/story.js\");\n/* harmony import */ var _avo_image_asset__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/image-asset */ \"./src/avo/image-asset.js\");\n/* harmony import */ var _avo_entity_types_hero__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/entity/types/hero */ \"./src/avo/entity/types/hero.js\");\n/* harmony import */ var _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/entity/types/wall */ \"./src/avo/entity/types/wall.js\");\n/* harmony import */ var _avo_entity_types_ball__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/entity/types/ball */ \"./src/avo/entity/types/ball.js\");\n/* harmony import */ var _avo_entity_types_enemy__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! @avo/entity/types/enemy */ \"./src/avo/entity/types/enemy.js\");\n/* harmony import */ var _avo_rule_types_zelda_controls__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! @avo/rule/types/zelda-controls */ \"./src/avo/rule/types/zelda-controls.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n\n\n\n\nvar ExampleStory = /*#__PURE__*/function (_Story) {\n  _inherits(ExampleStory, _Story);\n  var _super = _createSuper(ExampleStory);\n  function ExampleStory(app) {\n    _classCallCheck(this, ExampleStory);\n    return _super.call(this, app);\n  }\n  _createClass(ExampleStory, [{\n    key: \"assets\",\n    get: function get() {\n      return {\n        \"hero-4dir\": new _avo_image_asset__WEBPACK_IMPORTED_MODULE_2__[\"default\"]('assets/avo-sprites-2022-05-samiel.png'),\n        \"hero-2dir\": new _avo_image_asset__WEBPACK_IMPORTED_MODULE_2__[\"default\"]('assets/avo-sprites-2022-10-samiel-2dir.png')\n      };\n    }\n  }, {\n    key: \"start\",\n    value: function start() {\n      _get(_getPrototypeOf(ExampleStory.prototype), \"start\", this).call(this);\n      this.generate_zelda_default();\n    }\n\n    /*\r\n    Default top-down adventure level.\r\n     */\n  }, {\n    key: \"generate_zelda_default\",\n    value: function generate_zelda_default() {\n      var app = this._app;\n      app.hero = app.addEntity(new _avo_entity_types_hero__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 11, 1));\n      app.camera.target = app.hero;\n      app.addRule(new _avo_rule_types_zelda_controls__WEBPACK_IMPORTED_MODULE_7__[\"default\"](app));\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 0, 0, 1, 23)); // West Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 22, 0, 1, 23)); // East Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 1, 0, 21, 1)); // North Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 1, 22, 21, 1)); // South Wall\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 3, 2, 3, 1));\n      app.addEntity(new _avo_entity_types_wall__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 3, 4, 3, 1));\n      app.addEntity(new _avo_entity_types_ball__WEBPACK_IMPORTED_MODULE_5__[\"default\"](app, 8, 6));\n      var enemy = app.addEntity(new _avo_entity_types_enemy__WEBPACK_IMPORTED_MODULE_6__[\"default\"](app, 4, 8));\n      enemy.rotation = -45 / 180 * Math.PI;\n    }\n  }]);\n  return ExampleStory;\n}(_story__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/types/example-story.js?");
 
 /***/ }),
